+++ tests/WpDateTimeTest.php
@@ -3,7 +3,7 @@
 namespace Rarst\WordPress\DateTime;
 
 use Brain\Monkey\Functions;
-use Brain\Monkey\WP\Filters;
+use Brain\Monkey\Filters;
 
 /**
  * @coversDefaultClass Rarst\WordPress\DateTime\WpDateTimeTrait
@@ -15,11 +15,11 @@ class WpDateTimeTest extends WpDateTimeTestCase {
 	 */
 	public function testCreateFromPost() {
 
-		Functions::when( 'get_post_field' )->alias( function ( $field, $post ) {
+		Functions\when( 'get_post_field' )->alias( function ( $field, $post ) {
 			return empty( $post->$field ) ? false : $post->$field;
 		} );
 
-		Functions::expect( 'get_option' )->with( 'timezone_string' )->andReturn( 'Europe/Kiev' );
+		Functions\expect( 'get_option' )->with( 'timezone_string' )->andReturn( 'Europe/Kiev' );
 
 		$this->assertFalse( WpDateTime::createFromPost( 'invalid post' ) );
 		$this->assertFalse( WpDateTime::createFromPost( new \stdClass(), 'invalid_field' ) );
@@ -57,21 +57,21 @@ class WpDateTimeTest extends WpDateTimeTestCase {
 
 		$wp_datetime = new WpDateTimeImmutable( null, new \DateTimeZone( 'Europe/Kiev' ) );
 
-		Functions::when( 'date_i18n' )->alias( function ( $format, $wp_timestamp ) use ( $wp_datetime ) {
+		Functions\when( 'date_i18n' )->alias( function ( $format, $wp_timestamp ) use ( $wp_datetime ) {
 
 			$unix_timestamp = $wp_timestamp - $wp_datetime->getOffset();
 
 			$this->assertEquals( $wp_datetime->getTimestamp(), $unix_timestamp );
 			$this->assertNotEquals( 'c', $format );
 			$this->assertNotEquals( 'r', $format );
-			$this->assertTrue( has_filter( 'pre_option_timezone_string', 'DateTimeZone->getName()', 10, 0 ) );
+			$this->assertTrue( has_filter( 'pre_option_timezone_string', 'DateTimeZone->getName()' ) );
 
 			return $wp_datetime->setTimestamp( $unix_timestamp )->format( $format );
 		} );
 
-		Filters::expectAdded( 'pre_option_timezone_string' )->twice();
+		Filters\expectAdded( 'pre_option_timezone_string' )->twice();
 		$this->assertEquals( $wp_datetime->format( DATE_W3C ), $wp_datetime->formatI18n( 'c' ) );
 		$this->assertEquals( $wp_datetime->format( DATE_RFC2822 ), $wp_datetime->formatI18n( 'r' ) );
-		$this->assertFalse( has_filter( 'pre_option_timezone_string', 'DateTimeZone->getName()', 10, 0 ) );
+		$this->assertFalse( has_filter( 'pre_option_timezone_string', 'DateTimeZone->getName()' ) );
 	}
 }