+++ tests/WpDateTimeTestCase.php
@@ -8,11 +8,11 @@ class WpDateTimeTestCase extends PHPUnit_Framework_TestCase {
 
 	protected function setUp() {
 		parent::setUp();
-		Monkey::setUpWP();
+		Monkey\setUp();
 	}
 
 	protected function tearDown() {
-		Monkey::tearDownWP();
+		Monkey\tearDown();
 		parent::tearDown();
 	}
 }