+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## Changed
+- Updated tests and Brain Monkey to version 2.1.
+
 ## Fixed
 - Instance `static` (not `self`) for correct behavior on subclassing.